Add return statement.
Adds `return` statement which can be used to return values from function
calls.

Also includes an idea that script execution can be broken with `return`
statement. If `return` statement is invoked in middle of script
execution, execution stops there and rest of the script is ignored. It
helps to create stuff like this:

    <%
    if request.is_ajax():
        response["Content-Type"] = "application/json";
        response.write("{\"foo\": \"bar\"}");
        return;
    end if;
    %>
    <html>
    <head>
        <title>blah blah</title>