Add Apache module.
This adds Apache module SAPI implementation which, when compiled can be
dynamically loaded in Apache 2.x as content handler. It then parses and
executes every file which has the Apache content type
`tempearly-script`.

It is capable of parsing both GET and POST data, parsing of multi-part
data is currently unimplemented just like in other SAPI implementations.

I observed random crashes of Apache when this module was compiled with
GCC. For some reason, it works better when compiled with Clang. Propably
related to garbage collection somehow and if so, it means it's a bug not
related to this Apache module. Apache modules are damn difficult to
debug though, so I can't be sure. More testing of the interpreter when
it's compiled with GCC is required.