Add `import` function.
This commit adds an global `import` function which can be used to import
shit from other files.

Example of it's usage:

    <%
        stuff_from_other_file = import("some_other_file.tly");

        response.write(stuff_from_other_file["variable_from_other_file"]);
    %>

Files can only be imported once. The resulting hash map is cached into
the interpreter and that cached version is returned on subsequent import
calls of the same file.