Add __getattr__ fallback.
Adds an __getattr__ fallback similiar in Python to objects/values. When
an attribute is looked for which does not exist, and the object has an
function named `__getattr__`, that function is invoked and value
returned by it used as the value of the attribute. If `__getattr__` does
not exist, an exception is thrown.

Secondly, modifies `import` function so that it returns an regular
object instead of hash map.