Alternative syntax proposal.
As discussed in issue #22, I am not happy with the current template
syntax. This commit modifies parser and examples for a new kind of
syntax, inspired by Jinja2 and Liquid templating engines.

New syntax in a nutshell:

    {% code block %}

    {{ escaped expression }}

    {! unescaped (dangerous) expression !}

    {# comment #}

Feel free to comment your opinion about this. If this completely sucks,
just close the PR.