Temporary fix to POST parsing problem.
This is *ugly*, non-optimal workaround to problem where POST data is not
parsed because character encoding is included in the MIME type (such as
`application/x-www-form-urlencoded; charset=UTF-8`).

It should be only temporary fix, we need to eventually implement
decoding and encoding various character encodings and add a real MIME
type parser.