Require C++11 compliant compiler.
I believe that the time has come for us to require C++11 compliant
compiler for this project. C++11 support (at least partial support for
it) can now be found in all major C++ compilers and to (ab)use features
given by the "new" version of C++, we need to require an compiler that
has support for it.