Store function arguments in Frame class.
It seems to be rather good idea to store function arguments in `Frame`
class so that they can be reached by garbage collector through the stack
frame.

Also fixes a bug where the function itself which already is stored in
the `Frame` is not being marked by the garbage collection process. Oops.