Redesign methods.
This commit continues our journey down the rabbit hole which seems to
contain some bad design choices. It's a rewrite of the function API,
which is now more sane as the slot values for example are now completely
optional and unbound instance methods for classes can be now detected
and treated specially.