Get rid of Value class.
This commit is not reviewable by any mortal being. You can test whether
it works however.

This commit changes the structure of the entire interpreter by ditching
the `Value` class and replacing it with garbage collected objects which
can be wrapped in an `Handle`. The reason for this is that the `Value`
class could not be wrapped in an handle and I did not want to modify it
so that it could. This caused problems with garbage collection since
`Value` class cannot be used to track which objects are in use and which
are not.