Add REPL support to Tempearly.
This is something I found useful while developing Tempearly: A pseudo
SAPI which adds REPL support to Tempearly. It's an command line user
interface where you type in Tempearly expressions, which are then parsed
and evaluated and finally result of the expression is displayed to user.
Exceptions are also printed with full stack trace for examination.